Sent Mail Box

+++ src/components/pages/ViewMail.js
@@ -1,37 +1,53 @@
 import React from "react";
-import {Button, Modal} from 'react-bootstrap';
+import { Button, Modal } from "react-bootstrap";
 import useHttp from "../hooks/use-http";
 import { useDispatch, useSelector } from "react-redux";
 import { mailActions } from "../Redux-Store/mail-slice";
 
 const ViewMail = (props) => {
   const { sendRequest } = useHttp();
-  const viewMail = useSelector((state) => state.mail.viewMail);
-
+  const viewMail = useSelector((state) =>
+    props.type === "sent"
+      ? state.mail.viewSentMail
+      : state.mail.viewReceivedMail
+  );
   const dispatch = useDispatch();
   const viewMailHandler = () => {
-    dispatch(mailActions.mailHandler());
+    dispatch(mailActions.viewMailHandle({ id: null }));
   };
 
   const deleteMailHandler = async () => {
-    let url;
-    if (props.type === "received") {
-      url = `https://ecommerce-auth-a598c-default-rtdb.firebaseio.com/rec${props.email}/${props.mail.id}.json`;
-      dispatch(mailActions.deleteReceivedMail(props.mail.id));
-    } else {
-      url = `https://ecommerce-auth-a598c-default-rtdb.firebaseio.com/sent${props.email}/${props.mail.id}.json`;
-      dispatch(mailActions.deleteSentMail(props.mail.id ));
+    if (viewMail) {
+      let url;
+      if (props.type === "received") {
+        url = `https://ecommerce-auth-a598c-default-rtdb.firebaseio.com/rec${props.email}/${viewMail.id}.json`;
+        dispatch(mailActions.deleteReceivedMail({ id: viewMail.id }));
+      } else {
+        url = `https://ecommerce-auth-a598c-default-rtdb.firebaseio.com/sent${props.email}/${viewMail.id}.json`;
+        dispatch(mailActions.deleteSentMail({ id: viewMail.id }));
+      }
+      // sendRequest({
+      //   url: url,
+      //   method: "DELETE",
+      // });
+      // // dispatch(mailActions.mailHandler());
+      // viewMailHandler();
+      try {
+        await sendRequest({
+          url: url,
+          method: "DELETE",
+        });
+        // If the delete operation is successful, reset the viewMail state
+        viewMailHandler();
+      } catch (error) {
+        // Handle error if necessary
+      } 
     }
-    sendRequest({
-      url: url,
-      method: "DELETE",
-    });
-    dispatch(mailActions.mailHandler());
   };
 
   return (
     <Modal
-      show={viewMail}
+      show={viewMail !== null}
       onHide={viewMailHandler}
       backdrop="static"
       keyboard={false}
@@ -42,15 +58,14 @@ const ViewMail = (props) => {
       <Modal.Header closeButton>
         <Modal.Title id="contained-modal-title-vcenter">Message</Modal.Title>
       </Modal.Header>
-      <Modal.Body>
-        {props.mail.body}
-      </Modal.Body>
+      <Modal.Body>{viewMail && viewMail.body}</Modal.Body>
       <Modal.Footer>
-        <Button variant="danger" onClick={deleteMailHandler}>Delete</Button>
+        <Button variant="danger" onClick={deleteMailHandler}>
+          Delete
+        </Button>
         <Button onClick={viewMailHandler}>Close</Button>
       </Modal.Footer>
     </Modal>
-  
   );
 };
